requirements.txt
flask
flask-login
flask-sqlalchemy
passlib
reportlab
python-dotenv
pyotp
